+++ dart/geodata/lib/src/ogcapi_features/model/ogc_feature_item.dart
@@ -27,7 +27,7 @@ class OGCFeatureItem extends FeatureItem with LinksAware {
     final data = meta['links'];
     if (data is Iterable<dynamic>) {
       try {
-        return Links.fromData(data);
+        return Links.fromJson(data);
       } on FormatException {
         // nop, could not parse, but then let empty links to be returned
       }