+++ dart/geodata/CHANGELOG.md
@@ -13,6 +13,7 @@ Note this is a breaking change.
 
 🛠 Refactoring:
 - [Feature properties and data maps as Map<String, dynamic> instead of Map<String, Object?> #147](https://github.com/navibyte/geospatial/issues/147)
+- [Refactor fromJson / toJson methods on model classes #150](https://github.com/navibyte/geospatial/issues/150)
 
 🛠 Maintenance:
 - [Lint rules without old strong mode, but with new "stricter type checks" #148](https://github.com/navibyte/geospatial/issues/148)